Added prefix p in pointer type variable

+++ DataAccessLayer/FabricItemService.h
@@ -9,7 +9,7 @@ class FabricItemService
 public:
 	DATAACCESSLAYER_API FabricItemService();
 	DATAACCESSLAYER_API ~FabricItemService();
-	DATAACCESSLAYER_API unsigned int insert(std::string strFabricName,std::string strShortcut,float fQuantity,float fUnitPrice,std::string strColor,std::string strFabricType);
+	DATAACCESSLAYER_API unsigned int insert(std::string strFabricName,std::string strShortcut,double fQuantity,double fUnitPrice,std::string strColor,std::string strFabricType);
 };
 
 #endif
